requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.4
et-xmlfile==1.1.0
idna==3.10
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
